requirements.txt
pynput
pyperclip
pystray
Pillow
tk
winshell